% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validation_errors}
\alias{validation_errors}
\title{Extract errors from a validation report}
\usage{
validation_errors(x)
}
\arguments{
\item{x}{A \code{qibla_validation} object.}
}
\value{
A data frame of error-level issues.
}
\description{
Extract errors from a validation report
}
