% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validation_warnings}
\alias{validation_warnings}
\title{Extract warnings from a validation report}
\usage{
validation_warnings(x)
}
\arguments{
\item{x}{A \code{qibla_validation} object.}
}
\value{
A data frame of warning-level issues.
}
\description{
Extract warnings from a validation report
}
